created Product class, wrote method for adding product to the db, wrote test ensuring this works
